Change variable to u32. I would say it's safe to say that an HD remaster will never request more than 2GB of RAM, much less 1.

+++ Core/HDRemaster.h
@@ -28,7 +28,7 @@ extern bool g_DoubleTextureCoordinates;
 
 struct HDRemaster {
 	std::string gameID;
-	u64 MemorySize;
+	u32 MemorySize;
 	bool DoubleTextureCoordinates;
 };
